Update tests to pass without l10n files.

+++ bedrock/base/tests/test_accepted_locales.py
@@ -41,7 +41,10 @@ class AcceptedLocalesTest(test_utils.TestCase):
         cls.PROD_LANGUAGES = settings.PROD_LANGUAGES
         cls.DEV_LANGUAGES = settings.DEV_LANGUAGES
         settings.PROD_LANGUAGES = ('en-US',)
-        shutil.move(cls.locale, cls.locale_bkp)
+        if os.path.exists(cls.locale):
+            shutil.move(cls.locale, cls.locale_bkp)
+        else:
+            cls.locale_bkp = None
         for loc in ('en-US', 'fr', 'templates'):
             os.makedirs(os.path.join(cls.locale, loc, 'LC_MESSAGES'))
         open(os.path.join(cls.locale, 'empty_file'), 'w').close()
@@ -54,7 +57,8 @@ class AcceptedLocalesTest(test_utils.TestCase):
         settings.PROD_LANGUAGES = cls.PROD_LANGUAGES
         settings.DEV_LANGUAGES = cls.DEV_LANGUAGES
         shutil.rmtree(cls.locale)
-        shutil.move(cls.locale_bkp, cls.locale)
+        if cls.locale_bkp:
+            shutil.move(cls.locale_bkp, cls.locale)
 
     def test_build_dev_languages(self):
         """Test that the list of dev locales is built properly.
